backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mist-automation-backend"
version = "1.0.0"
description = "FastAPI backend for Mist Automation & Backup application"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Proprietary"}
authors = [
    {name = "Mist Automation Team", email = "automation@example.com"}
]
keywords = ["mist", "automation", "backup", "fastapi", "workflow"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Framework :: FastAPI",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "motor>=3.6.0",
    "pymongo>=4.9,<4.10",
    "beanie>=1.27.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.12",
    "pyotp>=2.9.0",
    "qrcode>=8.0",
    "cryptography>=43.0.0",
    "mistapi>=0.13.2",
    "httpx>=0.28.0",
    "aiohttp>=3.10.10",
    "celery>=5.4.0",
    "redis>=5.2.0",
    "apscheduler>=3.10.4",
    "gitpython>=3.1.43",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "jinja2>=3.1.4",
    "structlog>=24.4.0",
    "python-json-logger>=2.0.7",
]

[project.optional-dependencies]
dev = [
    "black>=24.10.0",
    "ruff>=0.7.3",
    "mypy>=1.13.0",
    "flake8>=7.1.0",
    "bandit>=1.8.0",
]
test = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "pytest-httpx>=0.30.0",
]
all = [
    "mist-automation-backend[dev,test]",
]

[project.urls]
Homepage = "https://github.com/yourusername/mist-automation"
Documentation = "https://github.com/yourusername/mist-automation/docs"
Repository = "https://github.com/yourusername/mist-automation"
Issues = "https://github.com/yourusername/mist-automation/issues"

[project.scripts]
mist-automation = "app.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["tests*"]

[tool.black]
line-length = 120
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | venv
  | \.mypy_cache
  | \.pytest_cache
  | \.tox
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 120
target-version = "py310"
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    ".mypy_cache",
    ".pytest_cache",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py
"tests/*" = ["B011"]  # assert False banned

[tool.ruff.lint.isort]
known-first-party = ["app"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "motor.*",
    "beanie.*",
    "jose.*",
    "passlib.*",
    "pyotp.*",
    "qrcode.*",
    "mistapi.*",
    "celery.*",
    "apscheduler.*",
    "git.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv"]
skips = ["B101", "B601"]  # assert_used, paramiko_calls

[tool.bandit.assert_used]
skip_dirs = ["tests"]
